mcp_oauth/pyproject.toml
[project]
name = "mcp-oauth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.2.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
